# swagger.yaml
---
openapi: 3.0.0
info:
  title: 사용자 및 상품 관리 API
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 개발 서버
tags:
  - name: Users
    description: 회원 관련 엔드포인트
  - name: Items
    description: 상품 관련 엔드포인트
paths:
  /signup:
    post:
      tags:
        - Users
      summary: 회원가입
      description: 새로운 사용자를 등록합니다.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/signup_body"
        required: true
      responses:
        "201":
          description: 사용자가 성공적으로 생성됨
        "400":
          description: 잘못된 요청
        "409":
          description: 이미 사용 중인 이메일
  /login:
    post:
      tags:
        - Users
      summary: 로그인
      description: 사용자 로그인을 처리합니다.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/login_body"
        required: true
      responses:
        "200":
          description: 성공적으로 로그인됨
        "401":
          description: 유효하지 않은 로그인 정보
  /logout:
    post:
      tags:
        - Users
      summary: 로그아웃
      description: 현재 사용자를 로그아웃합니다.
      responses:
        "200":
          description: 로그아웃 성공
      security:
        - bearerAuth: []
  /user:
    get:
      tags:
        - Users
      summary: 사용자 조회
      description: 로그인한 사용자의 정보를 조회합니다.
      responses:
        "200":
          description: 사용자 정보 반환
        "404":
          description: 사용자 정보를 찾을 수 없음
      security:
        - bearerAuth: []
    put:
      tags:
        - Users
      summary: 회원 정보 수정
      description: 사용자 정보를 수정합니다.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/user_body"
        required: true
      responses:
        "200":
          description: 사용자 정보 업데이트 성공
        "401":
          description: 현재 비밀번호가 일치하지 않음
        "404":
          description: 사용자를 찾을 수 없음
      security:
        - bearerAuth: []
    delete:
      tags:
        - Users
      summary: 회원 탈퇴
      description: 사용자 계정을 삭제합니다.
      responses:
        "200":
          description: 회원 탈퇴 성공
        "404":
          description: 사용자를 찾을 수 없음
      security:
        - bearerAuth: []
  /product:
    post:
      tags:
        - Items
      summary: 상품 등록
      description: 새로운 상품을 등록합니다.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/product_body"
        required: true
      responses:
        "201":
          description: 상품이 성공적으로 등록됨
        "400":
          description: 잘못된 요청
        "500":
          description: 서버 오류
      security:
        - bearerAuth: []
  /products:
    get:
      tags:
        - Items
      summary: 상품 목록 조회
      description: 상품 목록을 조회합니다.
      parameters:
        - name: sort
          in: query
          description: 정렬 방식
          required: false
          style: form
          explode: true
          schema:
            type: string
            enum:
              - ASC
              - DESC
      responses:
        "200":
          description: 상품 목록 반환
        "204":
          description: 등록된 상품 없음
        "400":
          description: 잘못된 요청
        "500":
          description: 서버 오류
  /product/{itemId}:
    get:
      tags:
        - Items
      summary: 상품 상세 조회
      description: 특정 상품의 상세 정보를 조회합니다.
      parameters:
        - name: itemId
          in: path
          description: 상품 ID
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      responses:
        "200":
          description: 상품 상세 정보 반환
        "404":
          description: 상품이 존재하지 않음
        "400":
          description: 잘못된 요청
        "500":
          description: 서버 오류
    put:
      tags:
        - Items
      summary: 상품 정보 수정
      description: 상품 정보를 수정합니다.
      parameters:
        - name: itemId
          in: path
          description: 상품 ID
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/product_itemId_body"
        required: true
      responses:
        "200":
          description: 상품 정보 수정 성공
        "404":
          description: 상품이 존재하지 않거나 권한 없음
        "400":
          description: 잘못된 요청
        "500":
          description: 서버 오류
      security:
        - bearerAuth: []
    delete:
      tags:
        - Items
      summary: 상품 삭제
      description: 상품을 삭제합니다.
      parameters:
        - name: itemId
          in: path
          description: 상품 ID
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      responses:
        "200":
          description: 상품 삭제 성공
        "404":
          description: 상품이 존재하지 않거나 권한 없음
        "400":
          description: 잘못된 요청
        "500":
          description: 서버 오류
      security:
        - bearerAuth: []
components:
  schemas:
    user_body:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
        name:
          type: string
    product_body:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
    login_body:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    product_itemId_body:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        status:
          type: string
    signup_body:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        confirmPassword:
          type: string
        name:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT